requirements.txt
aiomysql==0.2.0
aiosqlite==0.19.0
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
asyncpg==0.28.0
click==8.1.7
colorama==0.4.6
executing==2.0.1
fastapi==0.103.1
greenlet==3.0.3
h11==0.14.0
icecream==2.1.3
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
psycopg==3.1.10
psycopg-binary==3.1.10
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.17.2
PyMySQL==1.1.0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.23.2

requests~=2.31.0
aiogram==3.3.0